pyproject.toml
# SPDX-FileCopyrightText: 2022 Agathe Porte <microjoe@microjoe.org>
#
# SPDX-License-Identifier: MIT

[tool.poetry]
name = "lospec-palette-scrapper"
version = "0.1.0"
description = "A web scrapper for the Lospec color palettes."
authors = ["Agathe Porte <microjoe@microjoe.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.8.1"
SQLAlchemy = "^1.4.31"
alembic = "^1.7.7"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"